Expire old cached files every two hours.
This adds a cron trigger and a handler function ("scheduled()") for that trigger.
The handler deletes sufficiently-unused objects and their database rows.

Currently, this is deleting objects with BUCKET.delete(listOfKeys). I thought about
deleting objects one-at-a-time with some concurrency, but since the API provides
this handy bulk-delete method, I just used that. Even if it does the deletions
sequentially in the object store, it'll still save a bunch of network round trips.